Rename core→_core
