+++ anndata/readwrite/zarr.py
@@ -12,8 +12,8 @@ from pandas.api.types import is_categorical_dtype
 import numcodecs
 import zarr
 
-from ..core.anndata import AnnData
-from ..core.raw import Raw
+from .._core.anndata import AnnData
+from .._core.raw import Raw
 from ..compat import (
     _from_fixed_length_strings,
     _to_fixed_length_strings,
